+++ data/social/trade/exchange/grand_exchange.ifaces.toml
@@ -2,19 +2,26 @@
 id = 105
 type = "main_screen"
 components = {
-  buy_offer_1 = 30,
-  sell_offer_1 = 31,
-  buy_offer_2 = 46,
-  sell_offer_2 = 47,
-  buy_offer_3 = 62,
-  sell_offer_3 = 63,
-  buy_offer_4 = 81,
-  sell_offer_4 = 82,
-  buy_offer_5 = 100,
-  sell_offer_5 = 101,
-  buy_offer_6 = 119,
-  sell_offer_6 = 120,
+  view_offer_0 = 18,
+  buy_offer_0 = 30,
+  sell_offer_0 = 31,
+  view_offer_1 = 34,
+  buy_offer_1 = 46,
+  sell_offer_1 = 47,
+  view_offer_2 = 50,
+  buy_offer_2 = 62,
+  sell_offer_2 = 63,
+  view_offer_3 = 69,
+  buy_offer_3 = 81,
+  sell_offer_3 = 82,
+  view_offer_4 = 88,
+  buy_offer_4 = 100,
+  sell_offer_4 = 101,
+  view_offer_5 = 107,
+  buy_offer_5 = 119,
+  sell_offer_5 = 120,
   back = 127,
+  examine = 142,
   decrease_quantity = 157,
   increase_quantity = 159,
   add_1 = 162,
@@ -32,6 +39,7 @@ components = {
   offer_layer = 126,
   new_offer_layer = 156,
   buy_layer = 192,
+  choose_item = 194,
   sell_layer = 197,
   running_layer = 200,
   process_text = 201,
@@ -46,7 +54,7 @@ id = 106
 
 [stock_side]
 id = 107
-type = "inventory_tab"
+type = "overlay_tab"
 components = {
   items = {
     id = 18,
